Extend RAM memory component to suppport up to 4GB with 2 buses of 16 bits. Moved FileReader interface to internal components and renamed internal clementina sub-systems to "modules"

+++ pkg/components/cpu/instruction_actions_test.go
@@ -33,7 +33,7 @@ func evaluateRegisterValue(t *testing.T, cpu *Cpu65C02S, name string, value uint
 }
 
 func evaluateAddress(t *testing.T, cpu *Cpu65C02S, ram *memory.Ram, address uint16, expected uint8) {
-	value := ram.Peek(address)
+	value := ram.Peek(uint32(address))
 
 	if value != expected {
 		instruction := instructionSet.GetByOpCode(cpu.currentOpCode)
